Extract Measurement Version Dimension field into separate class
