+++ site-performance-tracker.php
@@ -60,6 +60,7 @@ require_once __DIR__ . '/php/src/class-plugin.php';
 require_once __DIR__ . '/php/src/class-settings.php';
 require_once __DIR__ . '/php/src/class-analyticstypesfield.php';
 require_once __DIR__ . '/php/src/class-analyticsidfield.php';
+require_once __DIR__ . '/php/src/class-measurementversiondimensionfield.php';
 require_once __DIR__ . '/php/helpers.php';
 
 /**
